# Try a few things to make our JS features faster