--- a/spec/features/editing_content/insert_inline_file_attachment_spec.rb
+++ b/spec/features/editing_content/insert_inline_file_attachment_spec.rb
@@ -19,7 +19,7 @@ RSpec.feature "Insert inline file attachment" do
     then_i_see_the_attachment_link_snippet_is_inserted
   end
 
-  scenario "Insert attachment without javascript", js: false do
+  scenario "without javascript" do
     given_there_is_an_edition_with_file_attachments
     when_i_go_to_edit_the_edition
     and_i_click_to_insert_a_file_attachment