--- a/spec/features/editing_content/insert_inline_file_attachment_spec.rb
+++ b/spec/features/editing_content/insert_inline_file_attachment_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 RSpec.feature "Insert inline file attachment" do
-  scenario "Insert attachment with javascript", js: true do
+  scenario "block snippet", js: true do
     given_there_is_an_edition_with_file_attachments
     when_i_go_to_edit_the_edition
     and_i_click_to_insert_a_file_attachment
@@ -10,7 +10,7 @@ RSpec.feature "Insert inline file attachment" do
     then_i_see_the_attachment_snippet_is_inserted
   end
 
-  scenario "Insert Attachment link with javascript", js: true do
+  scenario "link snippet", js: true do
     given_there_is_an_edition_with_file_attachments
     when_i_go_to_edit_the_edition
     and_i_click_to_insert_a_file_attachment