--- a/spec/features/editing_content/insert_inline_image_spec.rb
+++ b/spec/features/editing_content/insert_inline_image_spec.rb
@@ -9,7 +9,7 @@ RSpec.feature "Insert inline image" do
     then_i_see_the_snippet_is_inserted
   end
 
-  scenario "without javascript", js: false do
+  scenario "without javascript" do
     given_there_is_an_edition_with_images
     when_i_go_to_edit_the_edition
     and_i_click_to_insert_an_image