--- a/spec/features/editing_file_attachments/delete_file_attachment_spec.rb
+++ b/spec/features/editing_file_attachments/delete_file_attachment_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.feature "Delete a file attachment", js: true do
+RSpec.feature "Delete a file attachment" do
   scenario do
     given_there_is_an_edition_with_attachments
     when_i_insert_an_attachment