--- a/spec/features/editing_file_attachments/replace_file_attachment_file_spec.rb
+++ b/spec/features/editing_file_attachments/replace_file_attachment_file_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.feature "Replace a file attachment file", js: true do
+RSpec.feature "Replace a file attachment file" do
   scenario do
     given_there_is_an_edition_with_an_attachment
     when_i_click_to_insert_an_attachment