--- a/spec/features/editing_images/delete_image_spec.rb
+++ b/spec/features/editing_images/delete_image_spec.rb
@@ -46,8 +46,8 @@ RSpec.feature "Delete an image" do
   end
 
   def then_i_see_the_image_is_gone
-    expect(all("#image-#{@image_revision.image_id}").count).to be_zero
     expect(page).to have_content(I18n.t!("images.index.flashes.deleted", file: @image_revision.filename))
+    expect(page).to_not have_selector("#image-#{@image_revision.image_id}")
   end
 
   def and_i_see_the_timeline_entry