--- a/spec/features/editing_images/delete_image_spec.rb
+++ b/spec/features/editing_images/delete_image_spec.rb
@@ -9,7 +9,7 @@ RSpec.feature "Delete an image" do
     and_i_see_the_timeline_entry
   end
 
-  scenario "inline image", js: true do
+  scenario "inline image" do
     given_there_is_an_edition_with_images
     when_i_insert_an_inline_image
     and_i_delete_the_non_lead_image