Please see the commits for more details. We can always drop the last
commit if we're preferring slightly greater coverage for the modal.

## Example of an 'action override'
https://github.com/alphagov/content-publisher/blob/0950e16de6bddf4d67c694275b721799d75a2133/app/views/images/_image.html.erb#L35

## Before

    govuk-docker-run bundle exec rspec  1.63s user 0.28s system 1% cpu 2:48.54 total

## After

    govuk-docker-run bundle exec rspec  1.65s user 0.29s system 1% cpu 2:38.21 total