@kevindew 2020-02-07
Last commit has a typo in the description.

the 10 second benchmark looks good 👍 It does seem a bit of a shame to stop working with attachments through the user interface a user would as it seems to miss one of the components that is part of the integration test. 

Trying it myself made for benchmarks of that particular change made me question whether the time difference was worth it. On my local machine I'm seeing the feature tests taking ~2 seconds longer with them:

Before: 
```
➜  content-publisher git:(7c47e635) rspec spec/features
....................................................................

Finished in 44.17 seconds (files took 4.72 seconds to load)
68 examples, 0 failures
```

After:
```
➜  content-publisher git:(faster-js-features) rspec spec/features
....................................................................

Finished in 42.95 seconds (files took 4.74 seconds to load)
68 examples, 0 failures
```

Though on Docker it is ~4 seconds

Before:
```
➜  content-publisher git:(7c47e635) ✗ govuk-docker-run bundle exec rspec spec/features
docker-compose -f [...] run content-publisher-lite bundle exec rspec spec/features
Starting govuk-docker_redis_1 ... done
Starting govuk-docker_postgres_1 ... done
....................................................................

Finished in 1 minute 16.48 seconds (files took 12.31 seconds to load)
68 examples, 0 failures
```

After:
```
➜  content-publisher git:(faster-js-features) govuk-docker-run bundle exec rspec spec/features
docker-compose -f [...] run content-publisher-lite bundle exec rspec spec/features
Starting govuk-docker_redis_1 ... done
Starting govuk-docker_postgres_1 ... done
....................................................................

Finished in 1 minute 12.62 seconds (files took 12.36 seconds to load)
68 examples, 0 failures
```