@benthorner 2020-02-10
@kevindew ah, that's a shame. I was going to ask if those were average times, but I agree the argument for the last commit is pretty weak. The other commits should still be beneficial, so I've moved them to another PR: https://github.com/alphagov/content-publisher/pull/1752.